Move AgencyCard to components/index/cards
